seller serializers separated
